ㅎㅎ

+++ teatime/src/main/java/board/controller/StoryBoardController.java
@@ -29,7 +29,7 @@ import board.service.StoryBoardServiceImpl;
 
 @Controller
 public class StoryBoardController {
-	String user_id=null;
+	
 	@Autowired
 	StoryBoardServiceImpl service;
 	
@@ -44,9 +44,7 @@ public class StoryBoardController {
 	
 	//전체조회
 	@RequestMapping("teatime.storyBoardList")
-    public ModelAndView getStoryList(@RequestParam(value="pageNum", required=false, defaultValue="1") int pageNum,HttpSession session) {
-		user_id = (String)session.getAttribute("sessionId");
-		System.out.println("세션아이디:::::::::::"+user_id);
+    public ModelAndView getStoryList(@RequestParam(value="pageNum", required=false, defaultValue="1") int pageNum) {
 		ModelAndView mav = new ModelAndView();
 		List<StoryBoardDTO> list = service.getStoryBoardList(pageNum);
 		int pageNumCount = service.getPageNum();
@@ -98,7 +96,8 @@ public class StoryBoardController {
 	
 	//게시글 추가 후 게시판 전체조회로
 	@RequestMapping(value="teatime.storyBoardInsert",method=RequestMethod.POST)
-	public String setStoryInsert(StoryBoardDTO storyDTO){
+	public String setStoryInsert(StoryBoardDTO storyDTO,HttpSession session){
+		String user_id = (String)session.getAttribute("sessionId");
 		if(user_id!=null){
 			storyDTO.setUser_id(user_id);
 			String url = service.insertStoryBoard(storyDTO);		
@@ -111,10 +110,10 @@ public class StoryBoardController {
 	//게시글 수정
 	@RequestMapping(value="teatime.storyBoardModify",method=RequestMethod.GET)
 	public ModelAndView moveStoryModify(int idx, HttpSession session){
-		
+		String user_id = (String)session.getAttribute("sessionId");
 		ModelAndView mav = new ModelAndView();
 		StoryBoardDTO storyDTO = service.getDetailStoryBoard(idx);
-		storyDTO.setUser_id((String)session.getAttribute("sessionId"));
+		storyDTO.setUser_id(user_id);
 		mav.addObject("modify", storyDTO);
 		mav.setViewName("storyboard/StoryBoardModify");
 		return mav;	
@@ -124,19 +123,22 @@ public class StoryBoardController {
 	}
 	//게시글 수정
 	@RequestMapping(value="teatime.storyBoardModify",method=RequestMethod.POST)
-	public String setModifyStoryBoard(StoryBoardDTO storyDTO){
-		
-		String url = service.modifyStoryBoard(storyDTO);
-		
+	public String setModifyStoryBoard(StoryBoardDTO storyDTO,HttpSession session){
+		String user_id = (String)session.getAttribute("sessionId");
+		if(user_id!=null){
+			String url = service.modifyStoryBoard(storyDTO);
+		}
 		return "redirect:teatime.storyBoardList";					
 	}
 	
 	//게시글 삭제
 	@RequestMapping("teatime.storyBoardDelet")	
 	public String setDeleteStoryBoard(int idx, StoryBoardDTO storyDTO, HttpSession session){
-		storyDTO.setUser_id((String)session.getAttribute("sessionId"));
-		String url = service.deleteStoryBoard(idx);
-		
+		String user_id = (String)session.getAttribute("sessionId");
+		storyDTO.setUser_id(user_id);
+		if(user_id!=null){
+			String url = service.deleteStoryBoard(idx);
+		}
 		return "redirect:teatime.storyBoardList";
 	}
 	//게시글 신고 
@@ -146,9 +148,8 @@ public class StoryBoardController {
 	}
 			
 	@RequestMapping(value="teatime.storyBoardReport",method=RequestMethod.POST)
-	public String reportStoryBoard(int idx, String content){
-		//로그인 구현되면 해야함 임의로 정의
-		/*String user_id ="kgh";*/
+	public String reportStoryBoard(int idx, String content,HttpSession session){
+		String user_id = (String)session.getAttribute("sessionId");
 		String tableName="board";
 		service.insertReport(idx,user_id,content,tableName);		
 		return "redirect:teatime.storyBoardDetail?idx="+idx;
@@ -158,7 +159,7 @@ public class StoryBoardController {
 		@ResponseBody
 		@RequestMapping("teatime.checkReport")
 		public int checkReport(int idx, HttpSession session){
-			/*String user_id=(String)session.getAttribute("sessionId");*/
+			String user_id = (String)session.getAttribute("sessionId");
 			String tableName="board";
 			return service.checkReport(idx,user_id,tableName);
 		}
@@ -167,39 +168,47 @@ public class StoryBoardController {
 
 		@RequestMapping(value="teatime.storyBoardCommentInsert",method=RequestMethod.POST)
 		public String setStoryCommentInsert(StoryBoardCommentDTO storyCommentDTO, HttpSession session){
-			storyCommentDTO.setUser_id((String)session.getAttribute("sessionId"));
+			String user_id = (String)session.getAttribute("sessionId");
+			storyCommentDTO.setUser_id(user_id);
 			System.out.println("implcecommentinsertidx="+storyCommentDTO.getBoard_idx());
-			String url = serviceComment.insertStoryBoardComment(storyCommentDTO);				
+			if(user_id!=null){
+				String url = serviceComment.insertStoryBoardComment(storyCommentDTO);
+			}
 			return "redirect:teatime.storyBoardDetail?idx="+storyCommentDTO.getBoard_idx();					
 		}
 			
 	
 		@RequestMapping(value="teatime.storyBoardCommentModify",method=RequestMethod.POST)
-		public String moveStoryCommentModify(StoryBoardCommentDTO storyCommentDTO){
+		public String moveStoryCommentModify(StoryBoardCommentDTO storyCommentDTO,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");
 			storyCommentDTO.setUser_id(user_id);
-			String url = serviceComment.modifyStoryBoardComment(storyCommentDTO);			
+			if(user_id!=null){
+				String url = serviceComment.modifyStoryBoardComment(storyCommentDTO);
+			}
 			return "redirect:teatime.storyBoardDetail?idx="+storyCommentDTO.getBoard_idx();		
 		}
 		
 		
 		@RequestMapping(value="teatime.storyBoardCommentDelete",method=RequestMethod.GET)
-		public String setStoryCommentDelete(int idx, int board_idx){
-				
-			String url = serviceComment.deleteStoryBoardComment(idx);
-						
-			return "redirect:teatime.storyBoardDetail?idx="+board_idx;	
-			//return "redirect:teatime.storyBoardList";
+		public String setStoryCommentDelete(int idx, int board_idx,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");
+			if(user_id!=null){
+				String url = serviceComment.deleteStoryBoardComment(idx);
+			}		
+			return "redirect:teatime.storyBoardDetail?idx="+board_idx;		
 				
 		}	
 			
 	
 		@RequestMapping(value="teatime.storyBoardReCommentInsert",method=RequestMethod.POST)
 		public String setStoryReCommentInsert(StoryBoardCommentDTO storyCommentDTO, HttpSession session){
-			storyCommentDTO.setUser_id((String)session.getAttribute("sessionId"));
+			String user_id = (String)session.getAttribute("sessionId");
+			storyCommentDTO.setUser_id(user_id);
 			int indent = storyCommentDTO.getIndent();
 			storyCommentDTO.setIndent(indent+1);
-			String url = serviceComment.insertStoryBoardReComment(storyCommentDTO);
-				
+			if(user_id!=null){
+				String url = serviceComment.insertStoryBoardReComment(storyCommentDTO);
+			}
 			return "redirect:teatime.storyBoardDetail?idx="+storyCommentDTO.getBoard_idx();					
 		}
 		
@@ -213,13 +222,14 @@ public class StoryBoardController {
 		//게시글 추천 추가 
 		@ResponseBody
 		@RequestMapping("teatime.storyBoardRecommend")
-		public Map<String,Object> setRecommend(int idx,String recommend){
-			//user_id 임의 설정
-			/*String user_id="kgh";*/
-			System.out.println("세션아이디 :::::::::"+user_id);
+		public Map<String,Object> setRecommend(int idx,String recommend,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");			
 			String tableName ="board";
-			service.checkRecommend(idx,user_id,recommend,tableName);
+			if(user_id!=null){
+				service.checkRecommend(idx,user_id,recommend,tableName);
+			}
 			Map<String,Object> map = service.selectRecommend(idx,tableName);
+			
 			return map;
 			
 		}
@@ -227,11 +237,12 @@ public class StoryBoardController {
 		//댓글의 추천 및 비추천
 		@ResponseBody
 		@RequestMapping("teatime.commentRecommend")
-		public Map<String,Object> setCommentRecommend(int idx, String recommend){
-			//user_id 임의 설정
-			/*String user_id ="test";*/
+		public Map<String,Object> setCommentRecommend(int idx, String recommend,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");
 			String tableName="comment";
-			service.checkRecommend(idx,user_id,recommend,tableName);	
+			if(user_id!=null){
+				service.checkRecommend(idx,user_id,recommend,tableName);	
+			}
 			Map<String,Object> map = service.selectRecommend(idx,tableName);			
 			return map;
 		}
@@ -243,28 +254,33 @@ public class StoryBoardController {
 		}
 				
 		@RequestMapping(value="teatime.commentReport",method=RequestMethod.POST)
-		public String commentReport(int idx, String content, int board_idx){
-			//로그인 구현되면 해야함 임의로 정의
-			/*String user_id ="kgh";*/
+		public String commentReport(int idx, String content, int board_idx,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");
 			String tableName="comment";
-			service.insertReport(idx,user_id,content,tableName);		
+			if(user_id!=null){
+				service.insertReport(idx,user_id,content,tableName);
+			}
 			return "redirect:teatime.storyBoardDetail?idx="+board_idx;
 		}
 		
 		//이미 신고한 게시물인지 확인하기
 		@ResponseBody
 		@RequestMapping("teatime.checkCommentReport")
-		public int checkCommentReport(int idx){
-			/*String user_id="kgh";*/
+		public int checkCommentReport(int idx,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");
 			String tableName="comment";
-			return service.checkReport(idx,user_id,tableName);
+			int result =  service.checkReport(idx,user_id,tableName);
+			return result;
 		}
-		
+		//게시글 작성자와 로그인한 계정 비교
 		@ResponseBody
 		@RequestMapping("teatime.checkWriter")
-		public int checkWriter(int idx){
-			System.out.println("ajax 세션"+user_id);
-			System.out.println(service.checkWriter(user_id,idx));
-			return service.checkWriter(user_id,idx);
+		public int checkWriter(int idx,HttpSession session){
+			String user_id = (String)session.getAttribute("sessionId");
+			
+			int result = service.checkWriter(user_id,idx);
+			return result;
 		}
+		
+		
 }
